requirements.txt
asgiref==3.4.1
backports.entry-points-selectable==1.1.1
certifi==2021.10.8
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.3
Django==4.0.4
django-cors-headers==3.11.0
djangorestframework==3.13.1
filelock==3.4.0
idna==3.3
itypes==1.2.0
Jinja2==3.1.1
MarkupSafe==2.1.1
platformdirs==2.4.0
psycopg2==2.9.3
pytz==2021.3
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
virtualenv==20.10.0
